11css //좋아요 및 페이징 관련 변경 clipController clipdao clipMapper mybatis Like vo 추가
reviewcontroller
sc_07
sc_07
sc_11
SearchSchedulecontroller

+++ GitTest/src/main/webapp/WEB-INF/views/SC_07place.jsp
@@ -69,13 +69,14 @@ html, body {
 
  
 <script type="text/javascript">
- 
+var login_id=${sessionScope.user_id};
+
 
 function pagingFormSubmit(currentPage) { //currentPage가 어디서 호출되어 온다
    var form=document.getElementById("pagingForm");
    var page=document.getElementById("page"); 
-   page.value=currentPage; 
-   form.submit();
+   page.value=currentPage;  
+   paging(page.value);
 }
 
 var check = 0;
@@ -136,8 +137,10 @@ function replyUpdate(REV_SQ, CONTENT_ID){
             html+= "<div class='media-body'>";
             html+= "<div class='media-heading'><h4>"+item.USER_ID;
             html+= "<small>"+item.UPD_YMD;
-            html+= "<a href='javascript:replyUpdateForm("+item.REV_SQ+","+item.CONTENT_ID+")'> 수정 </a> ";
-            html+= "<a href='#none' onclick='deleteReply("+item.REV_SQ+","+item.CONTENT_ID+")'> 삭제 </a>";
+            if(login_id == item.USER_ID){ 		
+        		html+= '<a href="javascript:replyUpdateForm('+item.REV_SQ+','+item.CONTENT_ID+')"> 수정 </a>';
+                html+= '<a href="#none" onclick="deleteReply('+item.REV_SQ+','+item.CONTENT_ID+')"> 삭제 </a>';
+                }
             html+= "</small></h4></div>";
             html+= "<pre>";
             html+= "<div id='"+item.REV_SQ+"'class='review_txt'>"+item.REV_TXT+"</div>"; 
@@ -179,8 +182,10 @@ function deleteReply(REV_SQ, CONTENT_ID){
 	                  html+= '<div class="media-body">';
 	                  html+= '<div class="media-heading"><h4>'+item.USER_ID;
 	                  html+= '<small>'+item.UPD_YMD; 
-	                  html+= '<a href="javascript:replyUpdateForm('+item.REV_SQ+','+item.CONTENT_ID+')"> 수정 </a>';
-	                  html+= '<a href="#none" onclick="deleteReply('+item.REV_SQ+','+item.CONTENT_ID+')"> 삭제 </a>';
+	                  if(login_id == item.USER_ID){ 		
+	              		  html+= '<a href="javascript:replyUpdateForm('+item.REV_SQ+','+item.CONTENT_ID+')"> 수정 </a>';
+	                      html+= '<a href="#none" onclick="deleteReply('+item.REV_SQ+','+item.CONTENT_ID+')"> 삭제 </a>';
+	                      }
 	                  html+= "</small></h4></div>";
 	                  html+= "<pre>";
 	                  html+= "<div id='"+item.REV_SQ+"'class='review_txt'>"+item.REV_TXT+"</div>"; 
@@ -193,12 +198,12 @@ function deleteReply(REV_SQ, CONTENT_ID){
 	                  html2+='<a href ="javascript:pagingFormSubmit('+onepageback+')">◀ </a>';
 	                  var j=0;
 	                  for( var i = data.navi.startPageGroup ; i <= data.navi.endPageGroup ; i++){
-	                     j=j+1;
-	                     console.log(j);
-	                      html2+='<a href="javascript:pagingFormSubmit('+data.navi.startPageGroup + 1+')">'+j+' </a>'; 
+	                	  var counter=data.navi.startPageGroup+j;
+	                      j=j+1;
+	                      html2+='<a href="javascript:pagingFormSubmit('+counter+')">'+j+' </a>'; 
 	                  } 
-	                  html2+='<a href="javascript:pagingFormSubmit('+data.navi.currentPage + 1+')">▶ </a>';
-	                  html2+='<a href="javascript:pagingFormSubmit('+data.navi.currentPage + data.navi.pagePerGroup+')">▶▶</a></div>';
+	                  html2+='<a href="javascript:pagingFormSubmit('+(data.navi.currentPage + 1)+')">▶ </a>';
+	                  html2+='<a href="javascript:pagingFormSubmit('+(data.navi.currentPage + data.navi.pagePerGroup)+')">▶▶</a></div>';
 	                  html2+='<form action="SC_07place" method="get" id="pagingForm">';
 	                  html2+='<input type="hidden" id="page" name="page">';
 	                  html2+='<input type="hidden" id="CONTENT_ID" name="CONTENT_ID" value="${contentid}">';
@@ -554,8 +559,10 @@ function f_Data(){
                   html+= '<div class="media-body">';
                   html+= '<div class="media-heading"><h4>'+item.USER_ID;
                   html+= '<small>'+item.UPD_YMD;
-                  html+= '<a href="javascript:replyUpdateForm('+item.REV_SQ+','+item.CONTENT_ID+')"> 수정 </a>';
-                  html+= '<a href="#none" onclick="deleteReply('+item.REV_SQ+','+item.CONTENT_ID+')"> 삭제 </a>';
+                  if(login_id == item.USER_ID){ 		
+              		  html+= '<a href="javascript:replyUpdateForm('+item.REV_SQ+','+item.CONTENT_ID+')"> 수정 </a>';
+                      html+= '<a href="#none" onclick="deleteReply('+item.REV_SQ+','+item.CONTENT_ID+')"> 삭제 </a>';
+                      }
                   html+= '</small></h4></div>';
                   html+= '<pre>';
                   html+= "<div id='"+item.REV_SQ+"'class='review_txt'>"+item.REV_TXT+"</div>"; 
@@ -568,13 +575,13 @@ function f_Data(){
                   html2+='<a href ="javascript:pagingFormSubmit('+onepageback+')">◀ </a>';
                   var j=0;
                   for( var i = data.navi.startPageGroup ; i <= data.navi.endPageGroup ; i++){
-                     j=j+1;
-                     console.log(j);
-                      html2+='<a href="javascript:pagingFormSubmit('+data.navi.startPageGroup + 1+')">'+j+' </a>'; 
+                	  var counter=data.navi.startPageGroup+j;
+                      j=j+1;
+                      html2+='<a href="javascript:pagingFormSubmit('+counter+')">'+j+' </a>'; 
                   } 
-                  html2+='<a href="javascript:pagingFormSubmit('+data.navi.currentPage + 1+')">▶ </a>';
-                  html2+='<a href="javascript:pagingFormSubmit('+data.navi.currentPage + data.navi.pagePerGroup+')">▶▶</a></div>';
-                  html2+='<form action="SC_07place" method="get" id="pagingForm">';
+                  html2+='<a href="javascript:pagingFormSubmit('+(data.navi.currentPage + 1)+')">▶ </a>';
+                  html2+='<a href="javascript:pagingFormSubmit('+(data.navi.currentPage + data.navi.pagePerGroup)+')">▶▶</a></div>';
+                  html2+='<form action="read_Review" method="post" id="pagingForm">';
                   html2+='<input type="hidden" id="page" name="page">';
                   html2+='<input type="hidden" id="CONTENT_ID" name="CONTENT_ID" value="${contentid}">';
                   html2+='<input type="hidden" id="CONTENT_TYPE_ID" name="CONTENT_TYPE_ID" value="${contypeid}">';
@@ -636,13 +643,143 @@ function f_Data(){
       });
       
    });
+ 
+ 
 
 function locationObj(){
     var offset = $("#reviewWrite").offset();
     $('html, body').animate({scrollTop : offset.top}, 400);
 }
 
+function paging(page) { 
+	$.ajax({ 
+        type : "post",
+        url : "read_Review",
+        data :{
+           CONTENT_ID : contentId,
+           page : page
+        },
+        
+        success : function(data){
+           
+           console.log(data);
+            var html ="";
+            var html2 ="";
+            $.each(data.rList,function(index,item){
+              console.log(item); 
+              html+= "<div class='media'>";
+              html+= "<a class='pull-left' href='#'>";
+              html+= '<img class="media-object" src="http://placehold.it/64x64" alt=""></a>';
+              html+= '<div class="media-body">';
+              html+= '<div class="media-heading"><h4>'+item.USER_ID;
+              html+= '<small>'+item.UPD_YMD;
+              if(login_id == item.USER_ID){ 		
+          		  html+= '<a href="javascript:replyUpdateForm('+item.REV_SQ+','+item.CONTENT_ID+')"> 수정 </a>';
+                  html+= '<a href="#none" onclick="deleteReply('+item.REV_SQ+','+item.CONTENT_ID+')"> 삭제 </a>';
+                  }
+              html+= '</small></h4></div>';
+              html+= '<pre>';
+              html+= "<div id='"+item.REV_SQ+"'class='review_txt'>"+item.REV_TXT+"</div>"; 
+  				html+= "</pre></div></div>";
+            })
+              html2+='<div align="center">';
+              var bigpageback=data.navi.currentPage - data.navi.pagePerGroup;
+              html2+='<a href ="javascript:pagingFormSubmit('+bigpageback +')">◀◀ </a>';
+              var onepageback=data.navi.currentPage -1
+              html2+='<a href ="javascript:pagingFormSubmit('+onepageback+')">◀ </a>';
+              var j=0;
+              for( var i = data.navi.startPageGroup ; i <= data.navi.endPageGroup ; i++){
+            	  var counter=data.navi.startPageGroup+j;
+                  j=j+1;
+                  html2+='<a href="javascript:pagingFormSubmit('+counter+')">'+j+' </a>'; 
+              } 
+              html2+='<a href="javascript:pagingFormSubmit('+(data.navi.currentPage + 1)+')">▶ </a>';
+              html2+='<a href="javascript:pagingFormSubmit('+(data.navi.currentPage + data.navi.pagePerGroup)+')">▶▶</a></div>';
+              html2+='<form action="read_Review" method="post" id="pagingForm">';
+              html2+='<input type="hidden" id="page" name="page">';
+              html2+='<input type="hidden" id="CONTENT_ID" name="CONTENT_ID" value="${contentid}">';
+              html2+='<input type="hidden" id="CONTENT_TYPE_ID" name="CONTENT_TYPE_ID" value="${contypeid}">';
+              html2+='</form></div>'; 
+            
+            $("#review").html(html);
+            $("#paging").html(html2);
+            $('#reviewWrite').val('');      
+        },
+        error : function(e){
+           console.log(e);
+        }   
+     });
+};
 
+	
+
+
+
+$(function(){	
+	 
+
+	
+$.ajax({
+    
+    type : "post",
+    url : "read_Review",
+    data :{
+       CONTENT_ID : contentId            
+    },
+    
+    success : function(data){
+       
+       console.log(data);
+        var html ="";
+        var html2 ="";
+        $.each(data.rList,function(index,item){
+          console.log(item);
+        
+          html+= "<div class='media'>";
+          html+= "<a class='pull-left' href='#'>";
+          html+= '<img class="media-object" src="http://placehold.it/64x64" alt=""></a>';
+          html+= '<div class="media-body">';
+          html+= '<div class="media-heading"><h4>'+item.USER_ID;
+          html+= '<small>'+item.UPD_YMD;
+          if(login_id == item.USER_ID){ 		
+  		  html+= '<a href="javascript:replyUpdateForm('+item.REV_SQ+','+item.CONTENT_ID+')"> 수정 </a>';
+          html+= '<a href="#none" onclick="deleteReply('+item.REV_SQ+','+item.CONTENT_ID+')"> 삭제 </a>';
+          }
+          html+= '</small></h4></div>';
+          html+= '<pre>';
+          html+= "<div id='"+item.REV_SQ+"'class='review_txt'>"+item.REV_TXT+"</div>"; 
+				html+= "</pre></div></div>";
+        });
+          html2+='<div align="center">';
+          var bigpageback=data.navi.currentPage - data.navi.pagePerGroup;
+          html2+='<a href ="javascript:pagingFormSubmit('+bigpageback +')">◀◀ </a>';
+          var onepageback=data.navi.currentPage -1
+          html2+='<a href ="javascript:pagingFormSubmit('+onepageback+')">◀ </a>';
+          var j=0;
+          console.log(data.navi.endPageGroup);
+          for( var i = data.navi.startPageGroup ; i <= data.navi.endPageGroup ; i++){
+        	  var counter=data.navi.startPageGroup+j;
+        	  console.log(data.navi.startPageGroup);
+              j=j+1;
+              html2+='<a href="javascript:pagingFormSubmit('+j+')">'+j+' </a>'; 
+          } 
+          html2+='<a href="javascript:pagingFormSubmit('+(data.navi.currentPage + 1)+')">▶ </a>';
+          html2+='<a href="javascript:pagingFormSubmit('+(data.navi.currentPage + data.navi.pagePerGroup)+')">▶▶</a></div>';
+          html2+='<form action="read_Review" method="post" id="pagingForm">';
+          html2+='<input type="hidden" id="page" name="page">';
+          html2+='<input type="hidden" id="CONTENT_ID" name="CONTENT_ID" value="${contentid}">';
+          html2+='<input type="hidden" id="CONTENT_TYPE_ID" name="CONTENT_TYPE_ID" value="${contypeid}">';
+          html2+='</form></div>'; 
+        
+        $("#review").html(html);
+        $("#paging").html(html2);
+        $('#reviewWrite').val('');      
+    },
+    error : function(e){
+       console.log(e);
+    }   
+ });
+});
 
 </script>
 
@@ -1058,7 +1195,7 @@ function locationObj(){
                         <div class="form-group">
                             <textarea class="form-control" id="reviewWrite" rows="3"></textarea>
                         </div>
-                        <a href="#" class="btn btn-primary" id="write">등록</a> 
+                        <a href="#" onclick="return false" class="btn btn-primary" id="write">등록</a> 
                        <!--  <button type="submit" class="btn btn-primary">Submit</button> -->
                     </form>
                 </div>
@@ -1069,7 +1206,7 @@ function locationObj(){
 
                 <!-- Comment -->
                 <div id="review">
-                <c:forEach var="reply" items="${rList}">
+                <%-- <c:forEach var="reply" items="${rList}">
                 <div class="media">
                     <a class="pull-left" href="#">
                         <img class="media-object" src="http://placehold.it/64x64" alt="">
@@ -1077,22 +1214,22 @@ function locationObj(){
                     <div class="media-body">
                         <div class="media-heading"><h4>${reply.USER_ID}
                             <small>${reply.UPD_YMD}</small>
-                            <!-- 아이디 비교 -->
-                            <%-- <c:if test="${ID == reply.USER_ID}"> --%>
+                            
+                            <c:if test="${sessionScope.user_id == reply.USER_ID}">
                      <small><a href="javascript:replyUpdateForm('${reply.REV_SQ}','${reply.CONTENT_ID}')">수정</a>
                      		<a href="#none" onclick="deleteReply('${reply.REV_SQ}','${reply.CONTENT_ID}')">삭제</a></small>
-                     <%-- </c:if> --%>
+                     	</c:if> 
                         </h4></div>
                         <pre>
                         <div id="${reply.REV_SQ}" class="review_txt">${reply.REV_TXT}</div>
                         </pre>
                     </div>
                 </div>
-            </c:forEach>
+            </c:forEach> --%>
             </div>
             <div id="paging">
             
-            <div align="center">
+            <%-- <div align="center">
                <a href ="javascript:pagingFormSubmit(
                   ${navi.currentPage - navi.pagePerGroup})">◀◀</a>
                   <a href ="javascript:pagingFormSubmit(${navi.currentPage -1})">◀</a>
@@ -1113,7 +1250,7 @@ function locationObj(){
             <input type="hidden" id="CONTENT_ID" name="CONTENT_ID" value="${contentid}">
             <input type="hidden" id="CONTENT_TYPE_ID" name="CONTENT_TYPE_ID" value="${contypeid}">
             
-            </form>            
+            </form> --%>        
             </div>
                 <!-- Comment -->
             </div>
